split client.py
