+++ tda/client/__init__.py
@@ -1,0 +1,2 @@
+from .synchronous import Client
+from .asynchronous import AsyncClient
