+++ tda/client/base.py
@@ -13,7 +13,7 @@ import tda
 import time
 
 from tda.orders.generic import OrderBuilder
-from .utils import EnumEnforcer
+from ..utils import EnumEnforcer
 
 
 def get_logger():
@@ -1034,123 +1034,3 @@ class BaseClient(EnumEnforcer):
         path = '/v1/accounts/{}/watchlists/{}'.format(account_id, watchlist_id)
         return self._patch_request(path, watchlist_spec)
 
-class Client(BaseClient):
-    def _get_request(self, path, params):
-        dest = 'https://api.tdameritrade.com' + path
-
-        req_num = self._req_num()
-        self.logger.debug('Req {}: GET to {}, params={}'.format(
-            req_num, dest, json.dumps(params, indent=4)))
-
-        resp = self.session.get(dest, params=params)
-        self._log_response(resp, req_num)
-        tda.debug.register_redactions_from_response(resp)
-        return resp
-
-    def _post_request(self, path, data):
-        dest = 'https://api.tdameritrade.com' + path
-
-        req_num = self._req_num()
-        self.logger.debug('Req {}: POST to {}, json={}'.format(
-            req_num, dest, json.dumps(data, indent=4)))
-
-        resp = self.session.post(dest, json=data)
-        self._log_response(resp, req_num)
-        tda.debug.register_redactions_from_response(resp)
-        return resp
-
-    def _put_request(self, path, data):
-        dest = 'https://api.tdameritrade.com' + path
-
-        req_num = self._req_num()
-        self.logger.debug('Req {}: PUT to {}, json={}'.format(
-            req_num, dest, json.dumps(data, indent=4)))
-
-        resp = self.session.put(dest, json=data)
-        self._log_response(resp, req_num)
-        tda.debug.register_redactions_from_response(resp)
-        return resp
-
-    def _patch_request(self, path, data):
-        dest = 'https://api.tdameritrade.com' + path
-
-        req_num = self._req_num()
-        self.logger.debug('Req {}: PATCH to {}, json={}'.format(
-            req_num, dest, json.dumps(data, indent=4)))
-
-        resp = self.session.patch(dest, json=data)
-        self._log_response(resp, req_num)
-        tda.debug.register_redactions_from_response(resp)
-        return resp
-
-    def _delete_request(self, path):
-        dest = 'https://api.tdameritrade.com' + path
-
-        req_num = self._req_num()
-        self.logger.debug('Req {}: DELETE to {}'.format(req_num, dest))
-
-        resp = self.session.delete(dest)
-        self._log_response(resp, req_num)
-        tda.debug.register_redactions_from_response(resp)
-        return resp
-
-class AsyncClient(BaseClient):
-
-    async def _get_request(self, path, params):
-        dest = 'https://api.tdameritrade.com' + path
-
-        req_num = self._req_num()
-        self.logger.debug('Req {}: GET to {}, params={}'.format(
-            req_num, dest, json.dumps(params, indent=4)))
-
-        resp = await self.session.get(dest, params=params)
-        self._log_response(resp, req_num)
-        tda.debug.register_redactions_from_response(resp)
-        return resp
-
-    async def _post_request(self, path, data):
-        dest = 'https://api.tdameritrade.com' + path
-
-        req_num = self._req_num()
-        self.logger.debug('Req {}: POST to {}, json={}'.format(
-            req_num, dest, json.dumps(data, indent=4)))
-
-        resp = await self.session.post(dest, json=data)
-        self._log_response(resp, req_num)
-        tda.debug.register_redactions_from_response(resp)
-        return resp
-
-    async def _put_request(self, path, data):
-        dest = 'https://api.tdameritrade.com' + path
-
-        req_num = self._req_num()
-        self.logger.debug('Req {}: PUT to {}, json={}'.format(
-            req_num, dest, json.dumps(data, indent=4)))
-
-        resp = await self.session.put(dest, json=data)
-        self._log_response(resp, req_num)
-        tda.debug.register_redactions_from_response(resp)
-        return resp
-
-    async def _patch_request(self, path, data):
-        dest = 'https://api.tdameritrade.com' + path
-
-        req_num = self._req_num()
-        self.logger.debug('Req {}: PATCH to {}, json={}'.format(
-            req_num, dest, json.dumps(data, indent=4)))
-
-        resp = await self.session.patch(dest, json=data)
-        self._log_response(resp, req_num)
-        tda.debug.register_redactions_from_response(resp)
-        return resp
-
-    async def _delete_request(self, path):
-        dest = 'https://api.tdameritrade.com' + path
-
-        req_num = self._req_num()
-        self.logger.debug('Req {}: DELETE to {}'.format(req_num, dest))
-
-        resp = await self.session.delete(dest)
-        self._log_response(resp, req_num)
-        tda.debug.register_redactions_from_response(resp)
-        return resp
