+++ tests/test_client.py
@@ -98,7 +98,7 @@ class _TestClient:
 
     # get_orders_by_path
 
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_vanilla(self):
         self.client.get_orders_by_path(ACCOUNT_ID)
         self.mock_session.get.assert_called_once_with(
@@ -108,7 +108,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_vanilla_str(self):
         self.client.get_orders_by_path(str(ACCOUNT_ID))
         self.mock_session.get.assert_called_once_with(
@@ -118,7 +118,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_from_not_datetime(self):
         with self.assertRaises(ValueError) as cm:
             self.client.get_orders_by_path(
@@ -128,7 +128,7 @@ class _TestClient:
                          "from_entered_datetime, got 'builtins.str'")
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_to_not_datetime(self):
         with self.assertRaises(ValueError) as cm:
             self.client.get_orders_by_path(
@@ -138,7 +138,7 @@ class _TestClient:
                          "to_entered_datetime, got 'builtins.str'")
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_max_results(self):
         self.client.get_orders_by_path(ACCOUNT_ID, max_results=100)
         self.mock_session.get.assert_called_once_with(
@@ -149,7 +149,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_from_entered_datetime(self):
         self.client.get_orders_by_path(
             ACCOUNT_ID, from_entered_datetime=EARLIER_DATETIME)
@@ -160,7 +160,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_to_entered_datetime(self):
         self.client.get_orders_by_path(
             ACCOUNT_ID, to_entered_datetime=EARLIER_DATETIME)
@@ -171,14 +171,14 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_status_and_statuses(self):
         self.assertRaises(ValueError, lambda: self.client.get_orders_by_path(
             ACCOUNT_ID, to_entered_datetime=EARLIER_DATETIME,
             status='EXPIRED', statuses=[self.client_class.Order.Status.FILLED]))
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_status(self):
         self.client.get_orders_by_path(
             ACCOUNT_ID, status=self.client_class.Order.Status.FILLED)
@@ -190,7 +190,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_status_unchecked(self):
         self.client.set_enforce_enums(False)
         self.client.get_orders_by_path(ACCOUNT_ID, status='FILLED')
@@ -202,7 +202,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_statuses(self):
         self.client.get_orders_by_path(
             ACCOUNT_ID, statuses=[
@@ -216,7 +216,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_statuses_scalar(self):
         self.client.get_orders_by_path(
             ACCOUNT_ID, statuses=self.client_class.Order.Status.FILLED)
@@ -228,7 +228,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_path_statuses_unchecked(self):
         self.client.set_enforce_enums(False)
         self.client.get_orders_by_path(
@@ -243,7 +243,7 @@ class _TestClient:
     # get_orders_by_query
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_query_vanilla(self):
         self.client.get_orders_by_query()
         self.mock_session.get.assert_called_once_with(
@@ -253,7 +253,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_query_max_results(self):
         self.client.get_orders_by_query(max_results=100)
         self.mock_session.get.assert_called_once_with(
@@ -264,7 +264,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_query_from_entered_datetime(self):
         self.client.get_orders_by_query(from_entered_datetime=EARLIER_DATETIME)
         self.mock_session.get.assert_called_once_with(
@@ -274,7 +274,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_query_to_entered_datetime(self):
         self.client.get_orders_by_query(to_entered_datetime=EARLIER_DATETIME)
         self.mock_session.get.assert_called_once_with(
@@ -284,7 +284,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_query_status_and_statuses(self):
         with self.assertRaises(
                 ValueError, msg='at most one of status or statuses may be set'):
@@ -295,7 +295,7 @@ class _TestClient:
                     self.client_class.Order.Status.EXPIRED])
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_query_status(self):
         self.client.get_orders_by_query(status=self.client_class.Order.Status.FILLED)
         self.mock_session.get.assert_called_once_with(
@@ -306,7 +306,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_query_status_unchecked(self):
         self.client.set_enforce_enums(False)
         self.client.get_orders_by_query(status='FILLED')
@@ -318,7 +318,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_query_statuses(self):
         self.client.get_orders_by_query(statuses=[
             self.client_class.Order.Status.FILLED,
@@ -331,7 +331,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_query_statuses_scalar(self):
         self.client.get_orders_by_query(statuses=self.client_class.Order.Status.FILLED)
         self.mock_session.get.assert_called_once_with(
@@ -342,7 +342,7 @@ class _TestClient:
             })
 
     
-    @patch('tda.client.datetime.datetime', mockdatetime)
+    @patch('tda.client.base.datetime.datetime', mockdatetime)
     def test_get_orders_by_query_statuses_unchecked(self):
         self.client.set_enforce_enums(False)
         self.client.get_orders_by_query(statuses=['FILLED', 'EXPIRED'])
